pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Package comparing two file and show difference"
authors = ["M 3 <mesto3@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
